readme file updated, form added for index file, test for math function

+++ SMath.php
@@ -1,7 +1,7 @@
 <?php
 
 class Math {
-    public static function Sum($argumen1, $argumen2): int {
+    public static function Sum($argumen1, $argumen2): float {
         return $argumen1 + $argumen2;
     }
 }
